new cli option: anaconda worker deregister all

+++ binstar_build_client/worker_commands/deregister.py
@@ -2,25 +2,28 @@ from __future__ import (print_function, unicode_literals, division,
     absolute_import)
 
 import os
-import yaml
 import logging
+import platform
+import yaml
+
 
 from binstar_client.utils import get_binstar
 
 from binstar_build_client import BinstarBuildAPI
 from binstar_build_client.worker.register import WorkerConfiguration
 
-log = logging.getLogger('bisntar.build')
+log = logging.getLogger('binstar.build')
 
 def main(args, context="worker"):
 
     bs = get_binstar(args, cls=BinstarBuildAPI)
-
-    wconfig = WorkerConfiguration.load(args.worker_id)
-    wconfig.deregister(bs)
-
-    os.unlink(wconfig.filename)
-    log.debug("Removed worker config {}".format(wconfig.filename))
+    if args.worker_id.lower() == 'all':
+        WorkerConfiguration.deregister_all(bs)
+    else:
+        wconfig = WorkerConfiguration.load(args.worker_id)
+        wconfig.deregister(bs)
+        os.unlink(wconfig.filename)
+        log.debug("Removed worker config {0}".format(wconfig.filename))
 
 
 def add_parser(subparsers, name='deregister',
@@ -31,10 +34,10 @@ def add_parser(subparsers, name='deregister',
     parser = subparsers.add_parser(name,
                                    help=description, description=description,
                                    epilog=epilog)
-    parser.add_argument('-l', '--list',
-                        help='List the workers registered by this user/machine and exit.',
-                        action='store_true')
     parser.add_argument('worker_id',
-                        help="Worker id (required if no --config arg")
+                        help="Worker id to deregister or \"all\" to " +\
+                             "deregister all workers registered by this " +\
+                             "hostname: {0}".format(platform.node()),
+                        )
     parser.set_defaults(main=default_func)
     return parser
